feat: even slimmer helapordo.h
